requirements.txt
rich>=12.6.0
dnspython>=2.0.0  # optional